Ajustes quitando ResponseObject

+++ src/main/java/com/bancolombia/arka_javadevops/services/PerfilService.java
@@ -3,24 +3,18 @@ package com.bancolombia.arka_javadevops.services;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.bancolombia.arka_javadevops.models.Perfil;
 import com.bancolombia.arka_javadevops.repositories.PerfilRepository;
 
+import lombok.RequiredArgsConstructor;
+
 @Service //Denota que la clase forma parte de la capa de servicion como un bean de Spring
+@RequiredArgsConstructor
 public class PerfilService {
 
-    @Autowired  //Indica a Spring que busque una implementacion para esta interfaz
-                //Como PerfilRepository extiende de CrudRepository y en esta clase
-                //solo se usan metodos de CrudRepository no es necesario tener una clase que 
-                //implemente PerfilRepository
     private final PerfilRepository perfilRepository;
-
-    public PerfilService(PerfilRepository perfilRepository) {
-        this.perfilRepository = perfilRepository;
-    }
 
     public List<Perfil> obtenerPerfiles(){
         return (List<Perfil>) perfilRepository.findAll();
@@ -34,13 +28,25 @@ public class PerfilService {
         return perfilRepository.save(perfil);
     }
 
-    public void eliminarPerfil(int idPerfil){
-        perfilRepository.deleteById(idPerfil);
+    public Perfil actualizarPerfil(int idPerfil, Perfil perfil){
+        Optional<Perfil> perfilEncontrado = obtenerPerfilPorId(idPerfil);
+        if(perfilEncontrado.isPresent()){
+            perfil.setIdPerfil(idPerfil);
+            return perfilRepository.save(perfil);
+        }
+        return null;
+    }    
+
+    public boolean eliminarPerfil(int idPerfil){
+        Optional<Perfil> perfilEncontrado = obtenerPerfilPorId(idPerfil);
+        if(perfilEncontrado.isPresent()){
+            perfilRepository.deleteById(idPerfil);
+            return true;
+        }
+        return false;
     }
 
-    public Perfil actualizarPerfil(Perfil perfil){
-        return perfilRepository.save(perfil);
-    }
+
